Modifies payee route, views, and updates user#show.
Changes payee#new to manually include user_id in params.
Modifies user#show to be an index of user's payees.
Modifies links on payee pages to reroute back to user, not to payees path.
Removes category as an attribute of payee.
Modifies payee controller to set user for payees upon creation.